pyproject.toml
[project]
name = "repl-nix-workspace"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "flask>=3.1.2",
    "google-generativeai>=0.8.6",
    "numpy>=2.4.0",
    "openai>=2.14.0",
    "pypdf2>=3.0.1",
    "python-dotenv>=1.2.1",
    "scikit-learn>=1.8.0",
    "werkzeug>=3.1.5",
]
